requirements.txt
google-generativeai
pinecone
langchain
langchain-community
pypdf
python-dotenv
streamlit
numpy<2
